Giant Virtue Buff * Makes Giant a better virtue.
## About The Pull Request

<!-- Describe your pull request. Avoid text walls, use concise bullet points for easier readability. Document every change, or this can delay review and even discourage maintainers from merging your PR. -->
As it is right now, Giant isn't great beyond the meme value of being big. It's pretty taxing HP-wise to ram through doors and I haven't ever really found a scenario that benefitted from being able to throw someone on a non-reinforced grab. I propose it be made a little more impactful by making it add some strength/con and reduce speed, as well as give them an edge in grapples.
## Testing Evidence

<!-- It's mandatory to test your PR. Provide images, clips or description of how you tested your changes where possible. -->

## Why It's Good For The Game

Size matters in a fight. I think it should matter in-game, too. Giant is a really cool trait and I think it's potential is wasted in its current state.